move jsonToAnalyses to utils.tsx
